add more tests (#48)
* add more tests

* try to hack

* mergeqqq

* update

* update pipeline

+++ .buildkite/pipeline.sh
@@ -1,6 +1,25 @@
 #!/usr/bin/env bash
 
-cat <<EOF | tee /dev/tty | buildkite-agent pipeline upload
+echo "BUILDKITE_BRANCH: $BUILDKITE_BRANCH"
+
+if [[ $BUILDKITE_BRANCH == gh-readonly-queue* ]]; then
+  cat <<EOF | tee /dev/tty | buildkite-agent pipeline upload
+steps:
+  - name: "main"
+    command: "./check.sh"
+    timeout_in_minutes: 20
+    agents:
+      queue: "default"
+  - name: "main"
+    command: "./check.sh"
+    timeout_in_minutes: 20
+    agents:
+      queue: "default"
+EOF
+
+else
+
+  cat <<EOF | tee /dev/tty | buildkite-agent pipeline upload
 steps:
   - name: "main"
     command: "./check.sh"
@@ -8,3 +27,5 @@ steps:
     agents:
       queue: "default"
 EOF
+
+fi
